# get_nth_prev..._inst...: Look ahead and carve