@Grazfather 2021-03-02
Obsoletes #615

When looking to test the accuracy of #615 I noticed that although much faster, the 'greedy' approach was very often incorrect.

While starting at the old implementation I came up with an optimization: Disassemble up to addr if we can (don't cap # of instructions). If we disassemble more than we need, than take the last n+1 and make sure that it ends with the target address.

While looking at this I discovered another easy optimization: Instead of checking all the instructions for validity, ONLY do it if we know that the instructions end where we want. This avoids having to check instructions that we know won't lead us to where we need to go.

Compare old to new:
## New
```
Looking back 1 instructions          63 function calls in 0.001 seconds
Looking back 2 instructions          86 function calls in 0.002 seconds
Looking back 3 instructions          136 function calls in 0.003 seconds
Looking back 4 instructions          166 function calls in 0.003 seconds
Looking back 5 instructions          202 function calls in 0.004 seconds
Looking back 6 instructions          239 function calls in 0.004 seconds
Looking back 7 instructions          262 function calls in 0.004 seconds
Looking back 8 instructions          285 function calls in 0.005 seconds
Looking back 1 instructions          61 function calls in 0.001 seconds
Looking back 2 instructions          91 function calls in 0.002 seconds
Looking back 3 instructions          138 function calls in 0.003 seconds
Looking back 4 instructions          169 function calls in 0.004 seconds
Looking back 5 instructions          206 function calls in 0.005 seconds
Looking back 6 instructions          236 function calls in 0.005 seconds
Looking back 7 instructions          259 function calls in 0.005 seconds
Looking back 8 instructions          309 function calls in 0.006 seconds
Looking back 1 instructions          63 function calls in 0.001 seconds
Looking back 2 instructions          99 function calls in 0.002 seconds
Looking back 3 instructions          136 function calls in 0.003 seconds
Looking back 4 instructions          159 function calls in 0.003 seconds
Looking back 5 instructions          182 function calls in 0.003 seconds
Looking back 6 instructions          212 function calls in 0.004 seconds
Looking back 7 instructions          249 function calls in 0.005 seconds
Looking back 8 instructions          271 function calls in 0.005 seconds
         271 function calls in 0.005 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    0.005    0.005 {built-in method builtins.exec}
        1    0.000    0.000    0.005    0.005 <string>:1(<module>)
        1    0.000    0.000    0.005    0.005 gef.py:1297(newnth)
       36    0.002    0.000    0.004    0.000 gef.py:1238(gdb_disassemble)
       66    0.001    0.000    0.001    0.000 {method 'split' of 'str' objects}
        2    0.000    0.000    0.000    0.000 {method 'disassemble' of 'gdb.Architecture' objects}
        1    0.000    0.000    0.000    0.000 {built-in method builtins.all}
       34    0.000    0.000    0.000    0.000 gef.py:583(__init__)
       34    0.000    0.000    0.000    0.000 {method 'format' of 'str' objects}
       10    0.000    0.000    0.000    0.000 gef.py:1328(<genexpr>)
        1    0.000    0.000    0.000    0.000 gef.py:1391(gef_current_instruction)
       34    0.000    0.000    0.000    0.000 {method 'rstrip' of 'str' objects}
       35    0.000    0.000    0.000    0.000 {built-in method builtins.len}
        1    0.000    0.000    0.000    0.000 gef.py:1380(gef_instruction_n)
        9    0.000    0.000    0.000    0.000 gef.py:593(is_valid)
        2    0.000    0.000    0.000    0.000 {built-in method _gdb.selected_frame}
        2    0.000    0.000    0.000    0.000 {method 'architecture' of 'gdb.Frame' objects}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}

```

## Old
```
Looking back 1 instructions          336 function calls in 0.007 seconds
Looking back 2 instructions          840 function calls in 0.017 seconds
Looking back 3 instructions          1605 function calls in 0.031 seconds
Looking back 4 instructions          2346 function calls in 0.042 seconds
Looking back 5 instructions          3376 function calls in 0.059 seconds
Looking back 6 instructions          4592 function calls in 0.082 seconds
Looking back 7 instructions          5874 function calls in 0.107 seconds
Looking back 8 instructions          7452 function calls in 0.124 seconds
Looking back 1 instructions          329 function calls in 0.006 seconds
Looking back 2 instructions          799 function calls in 0.013 seconds
Looking back 3 instructions          1451 function calls in 0.026 seconds
Looking back 4 instructions          2306 function calls in 0.045 seconds
Looking back 5 instructions          3329 function calls in 0.062 seconds
Looking back 6 instructions          4421 function calls in 0.079 seconds
Looking back 7 instructions          5863 function calls in 0.109 seconds
Looking back 8 instructions          7563 function calls in 0.132 seconds
Looking back 1 instructions          359 function calls in 0.006 seconds
Looking back 2 instructions          873 function calls in 0.015 seconds
Looking back 3 instructions          1565 function calls in 0.027 seconds
Looking back 4 instructions          2339 function calls in 0.041 seconds
Looking back 5 instructions          3147 function calls in 0.054 seconds
Looking back 6 instructions          4438 function calls in 0.076 seconds
Looking back 7 instructions          6011 function calls in 0.104 seconds
Looking back 8 instructions          7668 function calls in 0.131 seconds
         7668 function calls in 0.131 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    0.131    0.131 {built-in method builtins.exec}
        1    0.000    0.000    0.131    0.131 <string>:1(<module>)
        1    0.009    0.009    0.131    0.131 gef.py:1263(oldnth)
      902    0.048    0.000    0.097    0.000 gef.py:1238(gdb_disassemble)
       90    0.006    0.000    0.023    0.000 {built-in method builtins.any}
      790    0.012    0.000    0.017    0.000 gef.py:1286(<genexpr>)
     1573    0.013    0.000    0.013    0.000 {method 'split' of 'str' objects}
       91    0.009    0.000    0.009    0.000 {method 'disassemble' of 'gdb.Architecture' objects}
      811    0.007    0.000    0.007    0.000 {method 'format' of 'str' objects}
      811    0.007    0.000    0.007    0.000 gef.py:583(__init__)
      901    0.007    0.000    0.007    0.000 {built-in method builtins.len}
      811    0.006    0.000    0.006    0.000 {method 'rstrip' of 'str' objects}
      700    0.005    0.000    0.005    0.000 gef.py:593(is_valid)
       91    0.001    0.000    0.001    0.000 {built-in method _gdb.selected_frame}
       91    0.001    0.000    0.001    0.000 {method 'architecture' of 'gdb.Frame' objects}
        1    0.000    0.000    0.000    0.000 gef.py:1391(gef_current_instruction)
        1    0.000    0.000    0.000    0.000 gef.py:1380(gef_instruction_n)
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
```